11. Учебный проект: разделяй и властвуй
